requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cyclonedx-python-lib==3.1.3
idna==3.4
mypy-extensions==0.4.3
packageurl-python==0.10.4
pathspec==0.10.3
platformdirs==2.6.2
Pygments==2.14.0
requests==2.28.1
rich==13.0.1
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.13
yara-python==4.2.3
